backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary>=2.9.9
neon==0.1.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
openai>=1.12.0
better-auth>=0.0.1b11
docker>=7.0.0